--- a/Cargo.toml
+++ b/Cargo.toml
@@ -30,7 +30,7 @@ clap-cargo = { version = "0.10.0", features = ["cargo_metadata"] }
 ignore = "0.4.18"
 clap-verbosity-flag = "2.0.0"
 log = "0.4.17"
-git2 = { version = "0.17.0", default-features = false }
+gix = { version = "0.52.0", default-features = false, features = ["max-performance"] }
 tame-index = { version = "0.2", features = ["git", "sparse"] }
 human-panic = "1.0.3"
 bugreport = "0.5.0"