--- a/Cargo.toml
+++ b/Cargo.toml
@@ -30,8 +30,11 @@ clap-cargo = { version = "0.10.0", features = ["cargo_metadata"] }
 ignore = "0.4.18"
 clap-verbosity-flag = "2.0.0"
 log = "0.4.17"
-gix = { version = "0.52.0", default-features = false, features = ["max-performance"] }
-tame-index = { version = "0.2", features = ["git", "sparse"] }
+# Note that `tame-index` and `gix` must be upgraded in lock-step to retain the same `gix`
+# minor version. Otherwise, one will compile `gix` two times in different minor versions.
+gix = { version = "0.53.1", default-features = false, features = ["max-performance", "revision"] }
+tame-index = { version = "0.6.0", features = ["git", "sparse"] }
+
 human-panic = "1.0.3"
 bugreport = "0.5.0"
 itertools = "0.11.0"