--- a/src/rustdoc_gen.rs
+++ b/src/rustdoc_gen.rs
@@ -515,45 +515,38 @@ impl RustdocFromGitRevision {
         config: &mut GlobalConfig,
     ) -> anyhow::Result<Self> {
         config.shell_status("Cloning", rev)?;
-        let repo = git2::Repository::discover(source)?;
+        let repo = gix::discover(source)?;
 
-        let rev = repo.revparse_single(rev)?;
-        let rev_dir = target.join(rev.id().to_string());
+        let tree_id = repo.rev_parse_single(&*format!("{rev}^{{tree}}"))?;
+        let tree_dir = target.join(tree_id.to_string());
 
-        std::fs::create_dir_all(&rev_dir)?;
-        let tree = rev.peel_to_tree()?;
-        extract_tree(&repo, tree, &rev_dir)?;
+        std::fs::create_dir_all(&tree_dir)?;
+        extract_tree(tree_id, &tree_dir)?;
 
-        let path = RustdocFromProjectRoot::new(&rev_dir, target)?;
+        let path = RustdocFromProjectRoot::new(&tree_dir, target)?;
         Ok(Self { path })
     }
 }
 
-fn extract_tree(
-    repo: &git2::Repository,
-    tree: git2::Tree<'_>,
-    target: &std::path::Path,
-) -> anyhow::Result<()> {
-    for entry in tree.iter() {
-        match entry.kind() {
-            Some(git2::ObjectType::Tree) => {
-                let object = entry.to_object(repo);
-                let tree = object.and_then(|o| o.peel_to_tree());
-                if let Ok(tree) = tree {
-                    let path = target.join(bytes2str(entry.name_bytes()));
-                    std::fs::create_dir_all(&path)?;
-                    extract_tree(repo, tree, &path)?;
-                }
+fn extract_tree(tree: gix::Id<'_>, target: &std::path::Path) -> anyhow::Result<()> {
+    for entry in tree.object()?.try_into_tree()?.iter() {
+        let entry = entry?;
+        match entry.mode() {
+            gix::object::tree::EntryMode::Tree => {
+                let path = target.join(bytes2str(entry.filename()));
+                std::fs::create_dir_all(&path)?;
+                extract_tree(entry.id(), &path)?;
             }
-            Some(git2::ObjectType::Blob) => {
-                let object = entry.to_object(repo);
-                let blob = object.and_then(|o| o.peel_to_blob());
-                if let Ok(blob) = blob {
-                    let path = target.join(bytes2str(entry.name_bytes()));
-                    let existing = std::fs::read(&path).ok();
-                    if existing.as_deref() != Some(blob.content()) {
-                        std::fs::write(&path, blob.content())?;
-                    }
+            gix::object::tree::EntryMode::Blob | gix::object::tree::EntryMode::BlobExecutable => {
+                let blob = entry.object()?;
+                assert!(
+                    blob.kind.is_blob(),
+                    "we are not working on a corrupted repository"
+                );
+                let path = target.join(bytes2str(entry.filename()));
+                let existing = std::fs::read(&path).ok();
+                if existing.as_deref() != Some(&blob.data) {
+                    std::fs::write(&path, &blob.data)?;
                 }
             }
             _ => {}