Switch from `git2` to `gix` to increase performance and reduce compile times. `tame-index` also uses `gix`, which means that previously two different `git` implementations were used. Now the implementation is unified, reducing the amount of dependencies.

Please note that this also means that, to avoid duplication of `gix` in the tree, each time `tame-index` is upgraded, one should validate that its `gix` dependency matches up with the `gix` dependency used here.

#### Tasks

* [x] upgrade
* [x] wait for a new release of `tame-index`

#### Measurements

Before the change:

```
cargo build  834.74s user 31.81s system 873% cpu 1:39.23 total
cargo build --release  548.64s user 24.55s system 827% cpu 1:09.27 total
```

After the change:

```
cargo build  752.77s user 27.67s system 904% cpu 1:26.31 total
cargo build --release  505.19s user 20.98s system 892% cpu 58.938 total
```

#### Why `gix` indicates breaking changes basically by default

`gix` uses `cargo smart-release` to handle setting manifest versions based on conventional git commit comments, and to release interconnected crates in the correct order, based on need. There is also the notion of *safety-bumps*, which means that the downstream will indicate breaking changes if an upstream crate indicates a breaking change. This is because some crates may re-export their dependencies, adding their API surface in the process. However, this isn't always the case and generally, `gix` indicates breaking changes way more than it would have to. Wouldn't it be nice if breaking changes could be detected instead of assumed?

I think it would be a dream if `cargo semver-check` could somehow be integrated into the release-process of `gix` so that *safety-bumps* could be done based on actual need. Thus far I didn't look into this as it's an optimization, but it would definitely be very, very nice to have. And I am just mentioning this here in case it's interesting or even inspiring.