@obi1kenobi 2023-09-08
> I think it would be a dream if cargo semver-check could somehow be integrated into the release-process of gix so that safety-bumps could be done based on actual need. Thus far I didn't look into this as it's an optimization, but it would definitely be very, very nice to have.

That'd be great! Unfortunately `cargo-semver-checks` currently can't connect different crates' APIs together since the rustc/rustdoc mechanism for doing this reliably isn't available yet. There's interest from the rustdoc team to do this, I just don't think it's happened yet.

It might be possible to do by scanning the 3rd party crates directly to find which APIs had breaking changes, then separately scanning `gix` to find which 3rd party APIs it uses, and finally intersecting the two datasets. This would require some custom work that I unfortunately don't have cycles for, but I'd be happy to mentor contributors who might be interested in it!