@Byron 2023-09-08
> This is awesome, thank you! I especially love the `max-performance` feature name :)

I hope for a future version of `gix`, I will be introducing `ultra-performance` ;). For what's done here, I actually think `max-performance` isn't needed and the feature could just be removed. It's definitely your call - the trade here is performance for purity 😅.

> Would you mind adding a comment near the gix and tame-index requirements that mentions the need to keep them in sync?

Makes perfect sense and is now done.

Please note that I just finished some compile-time optimization work which comes to effect only once `tame-index` has upgraded as well. I will wait until that happened unless you want to rather see that happen in a separate PR. Please let me know.