@obi1kenobi 2023-09-08
> I hope for a future version of `gix`, I will be introducing `ultra-performance` ;). For what's done here, I actually think `max-performance` isn't needed and the feature could just be removed. It's definitely your call - the trade here is performance for purity 😅.

I trust your judgment. I haven't looked too closely at what the tradeoff is, and I'm happy to do what you think is best. Excited to get `ultra-performance` one day too! 🔥 

> Please note that I just finished some compile-time optimization work which comes to effect only once `tame-index` has upgraded as well. I will wait until that happened unless you want to rather see that happen in a separate PR. Please let me know.

Either way is totally fine. I'm probably not going to manage to make another release before I head to RustConf next week, so waiting is fine and probably won't make any difference. But again, your call, I'm equally happy to wait or merge this as-is.