@Byron 2023-09-09
Alright, then I will wait for `tame-index` so the switch to latest version of `gix` becomes possible. I'll keep you posted :).