@Byron 2023-09-22
The CI failure seems to [be unrelated](https://github.com/obi1kenobi/cargo-semver-checks/actions/runs/6276825419/job/17047584373?pr=531#step:9:13). Probably it works when restarted.