@Byron 2023-09-25
I am a bit puzzled about the reproducible CI error:

```
Error: failed to open crates.io git index

Caused by:
    Refusing to initialize the non-empty directory as '/home/runner/.cargo/registry/index/github.com-1ecc6299db9ec823'
```

Maybe this isn't the switch to `gix`, but rather, the switch to the latest `tame-index`? It seems to try to open the crates-index, and initialize it despite non-empty. Maybe it's related to the way the crates-index is created on CI?

I downgraded `tame-index` just to see if that makes CI pass. In the mean-time, do you @obi1kenobi  have suggestions on how to reproduce the CI failure locally?