@Byron 2023-09-25
The problem with CI I am running into is as follows: 

* the cache restores the `crates-index` directory in a form that contains a `.cache` directory, but no `.git` directory
* This causes `tame-index` to fail to open the `git` repository, which makes it think it can clone it
* That fails as the folder it tries to clone into isn't empty - it contains the `.cache` directory after all.

CC @Jake-Shadle who might have an idea if `tame-index` could possibly handle that *(and I am not suggesting it should)* and CC @obi1kenobi who might know more about this CI setup and how it could be adjusted to include a `.git` directory.

Thanks everyone for their help.