@obi1kenobi 2023-09-25
Our caching setup just uses the `Swatinem/rust-cache` action, so I'm not sure why a `.git` directory might not be present if it's expected to be there. I haven't had a chance to look into how Rust's `.cache` directory is expected to be structured, but I've deleted the GitHub Actions cache entry that the job was using, so we can see if the problem repeats even after a cache miss.