@obi1kenobi 2023-09-25
Ah with the wiped cache, some of the debugging code is causing the job to fail. Sorry about that!