@Byron 2023-09-25
No worries at all, thanks for your help! Maybe it works now :).