@obi1kenobi 2023-09-25
Ah, unfortunately it seems to have failed :(

I don't think there's anything particularly unusual about our caching setup, so I'm wondering why other projects haven't run into this issue as well. Perhaps it's just a matter of time / upgrading to newer versions until they do as well? If so, we should probably flag this with whoever we think is best-positioned to contain the problem before it starts affecting other projects.

Do you think this is more likely to be a `Swatinem/rust-cache` GitHub Action issue, a `tame-index` issue, or something fundamental in cargo / Rust?