@Byron 2023-09-25
Maybe it was a recent `rust-cache` update that now prevents `.git` directories to become a part of the cache, that would explain why it starts happening now. Maybe it's also related to older versions of `crates-index`/`tame-index` still being able to clone into an existing directory?

My suggestion is to remove the cache in this case and see if CI times are still acceptable or lower than the longest-running  job. Then the added time due to the index clone might not even be noticeable.